Migrate spotify to ts
